requirements.txt
django==3.2.10
langchain==0.0.184
faiss-cpu==1.7.3
google-generativeai==0.1.0
python-dotenv==0.19.2